+++ Demo/Application/DemoDropInViewController.swift
@@ -83,8 +83,8 @@ class DemoDropInViewController: DemoBaseViewController, DemoDropInViewDelegate {
                 return
             }
             
-            if result.isCancelled {
-                self.progressBlock?("Cancelled🎲")
+            if result.isCanceled {
+                self.progressBlock?("Canceled 🎲")
             } else {
                 self.didSelectApplePay = (result.paymentOptionType == .applePay)
                 self.progressBlock?("Ready for checkout...")